requirements.txt
pre-commit==4.2.0
nvitop
numpy
pandas
tqdm
argparse
wandb
librosa
matplotlib
scipy
audiomentations
pytorch-lightning
six
torchlibrosa
timm
urllib3==2.2.2
einops
nnAudio
pretty_midi
mir_eval
resampy
descript-audio-codec==1.0.0
audiofile==1.2.1
descript_audio_codec==1.0.0
descript_audiotools==0.7.2
einops==0.8.0
encodec==0.1.1
accelerate
diffusers
transformers
nnAudio==0.3.3
lmdb==1.6.2
munch
omegaconf
byol-pytorch
pillow
easydict
pickle5
autoflake
h5py
pedalboard
deepspeed
